refactor twin cli
