+++ testsuites/twin/src/lib.rs
@@ -1,1 +1,2 @@
-pub mod twin;
+pub mod runner;
+pub mod setup;
